+++ vortex-serde/src/file/mod.rs
@@ -3,4 +3,7 @@ mod footer;
 mod layouts;
 pub mod reader;
 
+#[cfg(test)]
+mod tests;
+
 pub const FULL_FOOTER_SIZE: usize = 20;
